Time: 115 ms (100%), Space: 77.8 MB (0%) - LeetHub

+++ 3462-maximum-sum-with-at-most-k-elements/3462-maximum-sum-with-at-most-k-elements.java
@@ -11,7 +11,7 @@ class Solution {
 
         list.sort(Collections.reverseOrder());
 
-        int maxSum = 0;
+        long maxSum = 0;
         for (int i = 0; i < k; i++) {
             maxSum += list.get(i);
         }
